+++ R/cpp11.R
@@ -16,8 +16,8 @@ c_bigfloat_to_double <- function(x) {
   .Call(`_bignum_c_bigfloat_to_double`, x)
 }
 
-c_bigfloat_format <- function(x) {
-  .Call(`_bignum_c_bigfloat_format`, x)
+c_bigfloat_format <- function(x, notation, digits, is_sigfig) {
+  .Call(`_bignum_c_bigfloat_format`, x, notation, digits, is_sigfig)
 }
 
 c_bigfloat_compare <- function(lhs, rhs, na_equal) {
@@ -196,8 +196,8 @@ c_biginteger_to_double <- function(x) {
   .Call(`_bignum_c_biginteger_to_double`, x)
 }
 
-c_biginteger_format <- function(x) {
-  .Call(`_bignum_c_biginteger_format`, x)
+c_biginteger_format <- function(x, notation) {
+  .Call(`_bignum_c_biginteger_format`, x, notation)
 }
 
 c_biginteger_compare <- function(lhs, rhs, na_equal) {