+++ src/biginteger_interface.cpp
@@ -1,6 +1,7 @@
 #include <cpp11.hpp>
 #include "operations.h"
 #include "compare.h"
+#include "format.h"
 #include "biginteger_vector.h"
 
 
@@ -82,8 +83,11 @@ cpp11::doubles c_biginteger_to_double(cpp11::strings x) {
  *  Other  *
  *---------*/
 [[cpp11::register]]
-cpp11::strings c_biginteger_format(cpp11::strings x) {
-  return biginteger_vector(x).format();
+cpp11::strings c_biginteger_format(cpp11::strings x, cpp11::strings notation) {
+  return format_biginteger_vector(
+    biginteger_vector(x),
+    format_notation(notation[0])
+  );
 }